refactor project, update readme and creating script for dev
