+++ backend/routes/join.py
@@ -1,14 +1,9 @@
 """
-Join endpoint routes.
+Join operation endpoints.
 """
 from fastapi import APIRouter, HTTPException, Request
 
-from backend.models import (
-    JoinResponse,
-    BridgeTableRequest,
-    BridgeTableResponse,
-    JoinWithBridgeRequest,
-)
+from backend.models import JoinResponse, JoinWithBridgeRequest
 from backend.services import SemanticJoinService
 
 
@@ -16,81 +11,6 @@ router = APIRouter(
     prefix="",
     tags=["join"],
 )
-
-
-@router.get("/")
-async def root():
-    """Root endpoint with API information."""
-    return {
-        "message": "Semantic Join API",
-        "version": "0.2.0",
-        "endpoints": {
-            "/health": "GET - Health check endpoint",
-            "/bridge-table": "POST - Create bridge table (supports RS-JP and CS-JP)",
-            "/join-from-bridge": "POST - Perform three-way join (R ⋈ bridge ⋈ S)",
-        },
-        "workflow": [
-            "1. POST /bridge-table → Get best matches using selected algorithm",
-            "   • join_method='row': RS-JP",
-            "   • join_method='column': CS-JP",
-            "2. POST /join-from-bridge → Three-way join: list_r ⋈ bridge_table ⋈ list_s"
-        ],
-        "algorithms": {
-            "RS-JP (row)": "Fast greedy algorithm using row-level co-occurrence",
-            "CS-JP (column)": "Global optimization considering semantic compatibility between matched pairs"
-        }
-    }
-
-
-@router.get("/health")
-async def health_check():
-    """Health check endpoint."""
-    return {"status": "healthy"}
-
-
-@router.post("/bridge-table", response_model=BridgeTableResponse)
-async def create_bridge_table(request_data: BridgeTableRequest, request: Request):
-    """
-    Create a bridge table using RS-JP or CS-JP algorithm.
-
-    This endpoint finds the best matches for each value in list_r from list_s
-    based on corpus co-occurrence data.
-
-    - RS-JP (row): Greedy algorithm, each R value independently picks best S
-    - CS-JP (column): Global optimization considering semantic compatibility
-
-    Args:
-        request_data: BridgeTableRequest containing:
-            - list_r: values to be matched
-            - list_s: candidate values
-            - join_method: "row" for RS-JP or "column" for CS-JP
-        request: FastAPI request object to access app state
-
-    Returns:
-        BridgeTableResponse with the best matches
-
-    Raises:
-        HTTPException: If the operation fails
-    """
-    try:
-        join_service: SemanticJoinService = request.app.state.join_service
-        bridge_table = join_service.create_bridge_table(
-            request_data.list_r,
-            request_data.list_s,
-            request_data.join_method
-        )
-
-        return BridgeTableResponse(
-            bridge_table=bridge_table,
-            total_r_values=len(request_data.list_r),
-            total_s_values=len(request_data.list_s),
-            total_candidates=len(bridge_table),
-        )
-    except Exception as e:
-        raise HTTPException(
-            status_code=500,
-            detail=f"Error creating bridge table: {str(e)}",
-        )
 
 
 @router.post("/join-from-bridge", response_model=JoinResponse)
