+++ backend/services/algorithms/__init__.py
@@ -1,0 +1,13 @@
+"""
+Semantic join algorithm implementations.
+"""
+from .base import BridgeAlgorithm
+from .rs_jp import RSJPAlgorithm
+from .cs_jp_lp import CSJPLPAlgorithm
+
+__all__ = [
+    "BridgeAlgorithm",
+    "RSJPAlgorithm",
+    "CSJPLPAlgorithm",
+]
+
