+++ backend/utils/__init__.py
@@ -1,0 +1,29 @@
+"""
+Utility modules for text normalization and corpus processing.
+"""
+from .normalization import (
+    NormalizationStrategy,
+    normalize_value,
+    set_normalization_strategy,
+)
+from .corpus_utils import (
+    CorpusParser,
+    extract_rows_from_wdc_dict,
+    stream_json_tables,
+    table_hash,
+    set_corpus_strategy,
+)
+
+__all__ = [
+    # Normalization
+    "NormalizationStrategy",
+    "normalize_value",
+    "set_normalization_strategy",
+    # Corpus utilities
+    "CorpusParser",
+    "extract_rows_from_wdc_dict",
+    "stream_json_tables",
+    "table_hash",
+    "set_corpus_strategy",
+]
+
